requirements.txt
-i https://pypi.doubanio.com/simple
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
mysqlclient==2.2.7
PyJWT==1.7.1
sqlparse==0.5.3
typing_extensions==4.13.2
